requirements.txt
# FastAPI framework and dependencies
fastapi==0.103.1
uvicorn==0.23.2
starlette==0.27.0
pydantic==2.3.0
email-validator==2.0.0.post2

# Database
sqlalchemy==2.0.21
alembic==1.12.0
aiosqlite==0.19.0

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Payments
stripe==6.5.0

# Email
sendgrid==6.10.0
jinja2==3.1.2

# HTTP client for payment services
aiohttp==3.8.5

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0

# Development tools
black==23.9.1
flake8==6.1.0
isort==5.12.0